- Removed I in Interface names ~ Renamed NetSubscribers to NetSubscriberList + Moved SMS implementations into proper package ~ renamed packages to smallcase
