+++ networklibrary/src/test/java/com/eis/communication/network/commands/AddPeerTest.java
@@ -1,9 +1,8 @@
-package com.eis.communication.network.Commands;
+package com.eis.communication.network.commands;
 
-import com.eis.communication.network.SMSNetworkManager;
+import com.eis.communication.network.smsnetwork.SMSNetworkManager;
 import com.eis.smslibrary.SMSPeer;
 
-import org.junit.Before;
 import org.junit.Test;
 
 import static org.junit.Assert.*;
