+++ networklibrary/src/test/java/com/eis/communication/network/commands/AddResourceTest.java
@@ -1,6 +1,6 @@
-package com.eis.communication.network.Commands;
+package com.eis.communication.network.commands;
 
-import com.eis.communication.network.SMSNetworkManager;
+import com.eis.communication.network.smsnetwork.SMSNetworkManager;
 
 import org.junit.Test;
